Radiation modules now explicitly compute SURFACE as well as atmospheric radiation.
No longer using SurfaceRadiation() process class. Logically it just does
not seem to make sense to divide radiation between atm and suface.
And the code is much cleaner this way. Plus, that is how RRTM and CAM
codes do it.

Along the way, also cleaned up and rationalized flux calculation in
greygas.py and nband.py.