Change how diagnostics are handled internally!
They are now just like input variables: regular attributes, but with a _diag_vars list
that is used to create the diagnostics dictionary on the fly.
This simplifies some of the code to create and maintain diagnostics.

Also change how diagnostics are passed up the process tree during timestepping to be more robust.

It is now automatic that a subprocess diagnostic like radiation.OLR is also accessible as parent.OLR