Call manual build script explicitly from setup.py again.
Cannot get the fully automated build process to work for now. The workaround is to tell setuptools
to include the compiled *.so objects in the build directory as package data.
This should let users install from source with
`python setup.py install`

Get rid of all the commented-out experimental stuff in setup.py. Revisit the build procedure later.