Add Emanual convection scheme.
Original Fortran source distribution from Kerry Emanual (CONVECT43c) is included with documentation.
We have a custom wrapper for this code that is built with f2py
and a new climlab process object available as climlab.convection.EmanuelConvection()

Seems to work well but need to add a few tests.

Also waiting for permission from K. Emanuel before making this public.