Change axis naming conventions.
axis_type is changed to just axis to play better with xgcm
points, bounds and delta are gone. Now each axis has a name axis and two xr.DataArray objects axis and axis_bounds (where axis is a string such as lat or depth).

delta is computed by np.diff of the axis_bounds axis wherever it is needed.